apps/backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.2
gunicorn==23.0.0
python-dotenv==1.0.1
